:bug: rearranged utils for server build
