--- src/blocks/Biometrics/biometrics.web.js
@@ -1,4 +1,0 @@
-export default {
-  isSupported: () => Promise.resolve(false),
-  authenticate: () => Promise.resolve(false),
-};
