+++ src/blocks/Biometrics/utils/auth.ios.js
@@ -1,4 +1,3 @@
-// @flow
 import auth from "react-native-touch-id";
 
 export default auth;
